dos: cache the .dsk into external flash (guest-XIP phase 3)
store_file_in_flash() on the image at core entry, behind DOS_XIP_CACHE
(default off -- nothing reads the pointer until phase 5, and a multi-second
"first boot of this image" stall for a pointer nobody dereferences is a
behaviour change this phase is specified not to have).

Measured under gwemu: ALLEYCAT.dsk (1,474,560 B) lands at 0x90000000 in
1,063 ms and the next boot hits the cache in 3 ms at the same address.
BATTLECHESS.dsk -- 66,060,288 B, the largest image in the tree -- also fits,
because 66 MB is 63 MiB against a 64 MiB part; it took 47 s. It does NOT
fit on a dual-boot install, where get_extflash_base() reserves
EXTFLASH_OFFSET at the bottom, and that case is handled as what it is: a
NULL return, a log line, and a guest that boots from SD exactly as before.

Called before dos_cpu_init() on purpose: MAX_OPEN_FILES is 8 and the DOS
core holds three FILE* for its whole run, so the cache's own handle has to
be transient (see the fopen entry in CLAUDE.md).

Also fixes a uint32 overflow in the flash cache's progress percentage that
caching a 63 MiB file exposed: total_bytes_processed * 100 wraps at
42,949,673 bytes, so the bar ran 0,25,50 then 0,25 again. No other core's
ROM had ever been big enough to reach it.

Submodule bump carries the INT 13h observer and the phase-3 results.

+++ Core/Src/gw_flash_alloc.c
@@ -391,7 +391,13 @@ static bool circular_flash_write(const char *file_path,
         total_bytes_processed += bytes_read;
 
         if (progress_cb) {
-            progress = (uint8_t)((total_bytes_processed * 100) / (*data_size));
+            /* 64-bit intermediate, and not for tidiness: total_bytes_processed
+             * is uint32, so `* 100` overflows at 42,949,673 bytes and the
+             * percentage wraps to 0 and climbs again. No ROM had ever been
+             * large enough to reach it (GBA tops out at 32 MB); a DOS .dsk is
+             * -- roms/dos/BATTLECHESS.dsk is 66,060,288 bytes, and under gwemu
+             * its progress ran 0,25,50 then 0,25 for exactly this reason. */
+            progress = (uint8_t)(((uint64_t)total_bytes_processed * 100) / (*data_size));
             progress_cb(*data_size, total_bytes_processed, progress);
         }
